requirements.txt
asgiref==3.6.0
Django==4.2
djangorestframework==3.14.0
mysqlclient==2.1.1
psycopg2==2.9.6
pytz==2023.3
sqlparse==0.4.3
asgiref==3.6.0
Django==4.2
djangorestframework==3.14.0
mysqlclient==2.1.1
psycopg2==2.9.6
pytz==2023.3
sqlparse==0.4.3
